pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["dunt <ntdu198@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
drf-spectacular = "^0.28.0"
django-debug-toolbar = "^4.4.6"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
coverage = "^7.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
